+++ crates/cheatnet/src/lib.rs
@@ -2,9 +2,6 @@ use state::CheatnetState;
 
 pub mod cheatcodes;
 pub mod constants;
-pub mod execution;
 pub mod forking;
-pub mod panic_data;
-pub mod rpc;
 pub mod runtime_extensions;
 pub mod state;
