+++ crates/cheatnet/src/runtime_extensions/call_to_blockifier_runtime_extension/execution/deprecated/syscalls.rs
@@ -12,7 +12,7 @@ use blockifier::execution::deprecated_syscalls::{
 use blockifier::execution::execution_utils::{write_maybe_relocatable, ReadOnlySegment};
 use blockifier::execution::hint_code;
 
-use crate::execution::deprecated::calls::execute_library_call;
+use crate::runtime_extensions::call_to_blockifier_runtime_extension::execution::deprecated::calls::execute_library_call;
 use cairo_felt::Felt252;
 use cairo_vm::hint_processor::builtin_hint_processor::builtin_hint_processor_definition::HintProcessorData;
 use cairo_vm::hint_processor::hint_processor_definition::{HintProcessorLogic, HintReference};
