+++ crates/cheatnet/tests/builtins/segment_arena.rs
@@ -1,6 +1,6 @@
 use crate::common::state::create_cheatnet_state;
 use crate::common::{deploy_contract, felt_selector_from_name, state::create_cached_state};
-use cheatnet::rpc::call_contract;
+use cheatnet::runtime_extensions::call_to_blockifier_runtime_extension::rpc::call_contract;
 
 #[test]
 fn segment_arena_simple() {
