+++ crates/cheatnet/tests/cheatcodes/deploy.rs
@@ -5,7 +5,9 @@ use cairo_felt::Felt252;
 use cairo_vm::vm::errors::hint_errors::HintError;
 use cheatnet::cheatcodes::deploy::{deploy, deploy_at};
 use cheatnet::cheatcodes::CheatcodeError;
-use cheatnet::rpc::{call_contract, CallContractFailure, CallContractResult};
+use cheatnet::runtime_extensions::call_to_blockifier_runtime_extension::rpc::{
+    call_contract, CallContractFailure, CallContractResult,
+};
 use conversions::felt252::FromShortString;
 use conversions::IntoConv;
 use runtime::EnhancedHintError;
