+++ crates/cheatnet/tests/cheatcodes/spoof.rs
@@ -10,7 +10,7 @@ use cairo_felt::Felt252;
 use cheatnet::cheatcodes::deploy::deploy;
 use cheatnet::state::CheatTarget;
 use cheatnet::{
-    rpc::call_contract,
+    runtime_extensions::call_to_blockifier_runtime_extension::rpc::call_contract,
     state::{BlockifierState, CheatnetState},
 };
 use conversions::{felt252::FromShortString, IntoConv};
