+++ crates/cheatnet/tests/cheatcodes/spy_events.rs
@@ -5,7 +5,7 @@ use cairo_lang_starknet::contract::starknet_keccak;
 use cairo_vm::hint_processor::hint_processor_utils::felt_to_usize;
 use cheatnet::cheatcodes::deploy::deploy;
 use cheatnet::cheatcodes::spy_events::{Event, SpyTarget};
-use cheatnet::rpc::call_contract;
+use cheatnet::runtime_extensions::call_to_blockifier_runtime_extension::rpc::call_contract;
 use conversions::felt252::FromShortString;
 use conversions::IntoConv;
 use std::vec;
