+++ crates/cheatnet/tests/cheatcodes/warp.rs
@@ -8,7 +8,7 @@ use crate::{
 };
 use cairo_felt::Felt252;
 use cheatnet::cheatcodes::deploy::deploy;
-use cheatnet::rpc::call_contract;
+use cheatnet::runtime_extensions::call_to_blockifier_runtime_extension::rpc::call_contract;
 use cheatnet::state::CheatTarget;
 use conversions::felt252::FromShortString;
 use conversions::IntoConv;
