+++ crates/cheatnet/tests/starknet/resources.rs
@@ -12,7 +12,7 @@ use crate::common::{
 };
 use cairo_felt::Felt252;
 use cheatnet::cheatcodes::deploy::deploy;
-use cheatnet::rpc::call_contract;
+use cheatnet::runtime_extensions::call_to_blockifier_runtime_extension::rpc::call_contract;
 
 // TODO (834): Verify values in this test
 #[test]
