+++ crates/cheatnet/tests/starknet/timestamp.rs
@@ -7,7 +7,7 @@ use crate::{
 };
 use cairo_felt::Felt252;
 use cheatnet::{
-    rpc::call_contract,
+    runtime_extensions::call_to_blockifier_runtime_extension::rpc::call_contract,
     state::{BlockifierState, CheatnetState},
 };
 use starknet_api::core::ContractAddress;
