+++ crates/forge-runner/src/running.rs
@@ -14,6 +14,7 @@ use blockifier::state::state_api::State;
 use cairo_felt::Felt252;
 use cairo_vm::serde::deserialize_program::HintParams;
 use cairo_vm::types::relocatable::Relocatable;
+use cheatnet::runtime_extensions::call_to_blockifier_runtime_extension::CallToBlockifierExtension;
 use cheatnet::runtime_extensions::cheatable_starknet_runtime_extension::CheatableStarknetRuntimeExtension;
 use cheatnet::runtime_extensions::forge_runtime_extension::{ForgeExtension, ForgeRuntime};
 use cheatnet::runtime_extensions::io_runtime_extension::IORuntimeExtension;
@@ -257,6 +258,13 @@ pub fn run_test_case(
         extended_runtime: cheatable_runtime,
     };
 
+    let call_to_blockifier_runtime = ExtendedRuntime {
+        extension: CallToBlockifierExtension {
+            lifetime: &PhantomData,
+        },
+        extended_runtime: io_runtime,
+    };
+
     let forge_extension = ForgeExtension {
         environment_variables: &runner_params.environment_variables,
         contracts: &runner_params.contracts,
@@ -264,7 +272,7 @@ pub fn run_test_case(
 
     let mut forge_runtime = ExtendedRuntime {
         extension: forge_extension,
-        extended_runtime: io_runtime,
+        extended_runtime: call_to_blockifier_runtime,
     };
 
     let latest_block_number = if let Some(ValidatedForkConfig {
@@ -378,9 +386,11 @@ fn get_all_execution_resources(runtime: &ForgeRuntime) -> ExecutionResources {
         .extended_runtime
         .extended_runtime
         .extended_runtime
+        .extended_runtime
         .hint_handler
         .resources;
     let cheatnet_used_resources = &runtime
+        .extended_runtime
         .extended_runtime
         .extended_runtime
         .extension
@@ -403,6 +413,7 @@ fn get_all_execution_resources(runtime: &ForgeRuntime) -> ExecutionResources {
 
 fn get_context<'a>(runtime: &'a ForgeRuntime) -> &'a EntryPointExecutionContext {
     runtime
+        .extended_runtime
         .extended_runtime
         .extended_runtime
         .extended_runtime
