+++ crates/forge-runner/src/sierra_casm_runner_gas.rs
@@ -71,6 +71,7 @@ where
             .extended_runtime
             .extended_runtime
             .extended_runtime
+            .extended_runtime
             .hint_handler,
         0,
         2,
